Plugin API Surface Refactor (#879)
* Refactored and cleaned up interfaces between codegen and cli packages

* Migrate js module to codegen crate.

* Removed feature flagging for now.

* Minor cleanup.

* Removed accidental downgrade of dependency in cli

* Fixed up a few comments/visibilities and moved uninitialized plugins to CLI crate.

* Fixed up vaious comments/docstrings

* Corrected linting violations.
