+++ crates/cli/Cargo.toml
@@ -23,9 +23,9 @@ wasmtime-wasi = { workspace = true }
 wasi-common = { workspace = true }
 walrus = "0.23.3"
 swc_core = { version = "10.7.0", features = [
-    "common_sourcemap",
-    "ecma_ast",
-    "ecma_parser",
+  "common_sourcemap",
+  "ecma_ast",
+  "ecma_parser",
 ] }
 wit-parser = "0.212.0"
 convert_case = "0.7.1"
