+++ crates/cli/src/main.rs
@@ -1,21 +1,19 @@
-mod bytecode;
 mod codegen;
 mod commands;
-mod js;
 mod js_config;
 mod option;
-mod plugins;
-mod wit;
+mod plugin;
 
-use crate::codegen::WitOptions;
 use crate::commands::{Cli, Command, EmitPluginCommandOpts};
 use anyhow::Result;
 use clap::Parser;
-use codegen::{CodeGenBuilder, CodeGenType};
+use codegen::js::JS;
+use codegen::{plugin::Plugin, wit::WitOptions, Generator, LinkingKind};
 use commands::CodegenOptionGroup;
-use js::JS;
 use js_config::JsConfig;
-use plugins::{Plugin, UninitializedPlugin};
+use plugin::{
+    CliPlugin, PluginKind, UninitializedPlugin, PLUGIN_MODULE, QUICKJS_PROVIDER_V2_MODULE,
+};
 use std::fs;
 use std::fs::File;
 use std::io::Write;
@@ -38,47 +36,65 @@ fn main() -> Result<()> {
             );
 
             let js = JS::from_file(&opts.input)?;
-            let mut builder = CodeGenBuilder::new();
-            builder
+
+            let mut generator = Generator::new();
+            generator
                 .wit_opts(WitOptions::from_tuple((
                     opts.wit.clone(),
                     opts.wit_world.clone(),
                 ))?)
-                .source_compression(!opts.no_source_compression);
+                .source_compression(!opts.no_source_compression)
+                .js_runtime_config(JsConfig::default().to_json()?);
 
-            let config = JsConfig::default();
-            let mut gen = if opts.dynamic {
-                builder.plugin(Plugin::V2);
-                builder.build(CodeGenType::Dynamic, config)?
+            if opts.dynamic {
+                generator
+                    .plugin(Plugin::new(QUICKJS_PROVIDER_V2_MODULE.to_vec()))
+                    .linking(LinkingKind::Dynamic)
+                    .linking_v2_plugin(true);
             } else {
-                builder.build(CodeGenType::Static, config)?
+                generator
+                    .plugin(Plugin::new(PLUGIN_MODULE.to_vec()))
+                    .linking(LinkingKind::Static)
+                    .linking_default_plugin(true);
             };
 
-            let wasm = gen.generate(&js)?;
+            let wasm = generator.generate(&js)?;
 
             fs::write(&opts.output, wasm)?;
             Ok(())
         }
         Command::Build(opts) => {
             let js = JS::from_file(&opts.input)?;
-            let codegen: CodegenOptionGroup = opts.codegen.clone().try_into()?;
-            let plugin = match codegen.plugin {
-                Some(path) => Plugin::new_user_plugin(&path)?,
-                None => Plugin::Default,
-            };
-            let js_opts = JsConfig::from_group_values(&plugin, opts.js.clone())?;
-            let mut builder = CodeGenBuilder::new();
-            builder
-                .wit_opts(codegen.wit)
-                .source_compression(codegen.source_compression)
-                .plugin(plugin);
-            let mut gen = if codegen.dynamic {
-                builder.build(CodeGenType::Dynamic, js_opts)?
-            } else {
-                builder.build(CodeGenType::Static, js_opts)?
+            let codegen_opts: CodegenOptionGroup = opts.codegen.clone().try_into()?;
+            let mut generator = Generator::new();
+
+            // Always assume the default plugin if no plugin is provided.
+            let cli_plugin = match &codegen_opts.plugin {
+                Some(path) => CliPlugin::new(Plugin::new_from_path(path)?, PluginKind::User),
+                None => CliPlugin::new(Plugin::new(PLUGIN_MODULE.to_vec()), PluginKind::Default),
             };
 
-            let wasm = gen.generate(&js)?;
+            let js_opts = JsConfig::from_group_values(&cli_plugin, opts.js.clone())?;
+
+            // Always link to the default plugin if no plugin is provided.
+            if codegen_opts.plugin.is_none() {
+                generator.linking_default_plugin(true);
+            }
+
+            // Configure the generator with the provided options.
+            generator
+                .plugin(cli_plugin.into_plugin())
+                .wit_opts(codegen_opts.wit)
+                .source_compression(!codegen_opts.source_compression)
+                .js_runtime_config(js_opts.to_json()?);
+
+            if codegen_opts.dynamic {
+                generator.linking(LinkingKind::Dynamic);
+            } else {
+                generator.linking(LinkingKind::Static);
+            };
+
+            let wasm = generator.generate(&js)?;
 
             fs::write(&opts.output, wasm)?;
             Ok(())
@@ -104,6 +120,6 @@ fn emit_plugin(opts: &EmitPluginCommandOpts) -> Result<()> {
         Some(path) => Box::new(File::create(path)?),
         _ => Box::new(std::io::stdout()),
     };
-    file.write_all(Plugin::Default.as_bytes())?;
+    file.write_all(PLUGIN_MODULE)?;
     Ok(())
 }
